Move LinkProcessingOptions into separate file
This makes future implementations easier where we will
reuse most of this code to parse Link headers.

Part of #35035
